+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/mtorre4580/back-meli/compare/v0.1.0...v0.2.0) (2019-04-18)
+
+
+### Features
+
+* add dnscache for performance ([7b6ffb0](https://github.com/mtorre4580/back-meli/commit/7b6ffb0))
+* add logger ([35c6a21](https://github.com/mtorre4580/back-meli/commit/35c6a21))
+* add middleware for status api /healthcheck ([9f9edae](https://github.com/mtorre4580/back-meli/commit/9f9edae))
+* add script to check performance autocannon ([30fcf4d](https://github.com/mtorre4580/back-meli/commit/30fcf4d))
+
+
+
 <a name="0.1.0"></a>
 # 0.1.0 (2019-04-18)